Tidy up.
