+++ src/core/models/game.ts
@@ -1,0 +1,17 @@
+export interface Game {
+  beats: StoryBeat[]
+}
+
+export enum BeatType {
+  Message,
+  Question,
+}
+
+export class StoryBeat {
+  constructor(private _beatType: BeatType) {
+  }
+
+  get beatType(): BeatType {
+    return this._beatType;
+  }
+}
